requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
pytest>=7.4.0
boto3>=1.34.0
moto[dynamodb]>=5.0.0
